Allow editing FrequencyType/Days and type FrequencyType as an enum
- ClientEditScheduledTask now supports FrequencyType and Days via the
  existing ModifyList convention. Modifying FrequencyType requires
  Days to be provided in the same request (frequency determines what
  Days means - weekly vs monthly); Days can still be edited alone.
- FrequencyType on ClientScheduledTask and ClientEditScheduledTask
  changed from int to ScheduledTaskFrequencyType, matching
  BaseScheduledTaskVm which already used the enum on the response
  side. No StringEnumConverter is registered anywhere in the
  solution, so Newtonsoft serializes/deserializes enums as plain
  numbers by default - verified with a throwaway console app that the
  JSON wire format is unchanged, so no frontend changes are needed.

+++ MyFinanceBackend/Data/AutomaticTaskRepository.cs
@@ -90,7 +90,7 @@ namespace MyFinanceBackend.Data
 				new SqlParameter(DatabaseConstants.PAR_DESCRIPTION, clientScheduledTask.Description),
 				new SqlParameter(DatabaseConstants.PAR_ACCOUNT_ID, clientScheduledTask.AccountId),
 				new SqlParameter(DatabaseConstants.PAR_IS_SPEND_TRX, clientScheduledTask.IsSpendTrx),
-				new SqlParameter(DatabaseConstants.PAR_PERIOD_TYPE_ID, clientScheduledTask.FrequencyType),
+				new SqlParameter(DatabaseConstants.PAR_PERIOD_TYPE_ID, (int)clientScheduledTask.FrequencyType),
 				new SqlParameter(DatabaseConstants.PAR_DAYS, sqlDays)
 			};
 
@@ -112,7 +112,7 @@ namespace MyFinanceBackend.Data
 				new SqlParameter(DatabaseConstants.PAR_DESCRIPTION, clientScheduledTask.Description),
 				new SqlParameter(DatabaseConstants.PAR_ACCOUNT_ID, clientScheduledTask.AccountId),
 				new SqlParameter(DatabaseConstants.PAR_TO_ACCOUNT_ID, clientScheduledTask.ToAccountId),
-				new SqlParameter(DatabaseConstants.PAR_PERIOD_TYPE_ID, clientScheduledTask.FrequencyType),
+				new SqlParameter(DatabaseConstants.PAR_PERIOD_TYPE_ID, (int)clientScheduledTask.FrequencyType),
 				new SqlParameter(DatabaseConstants.PAR_DAYS, sqlDays)
 			};
